MR - guest e2e tests done, minor modifications in guest service

See merge request ikt7645105/meeting-table!9

+++ backend/src/guest/guest.service.ts
@@ -1,5 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { HttpStatus } from '@nestjs/common/enums';
+import { HttpException } from '@nestjs/common/exceptions';
+import { PrismaClientKnownRequestError } from '@prisma/client/runtime';
+import { PrismaService } from './../prisma/prisma.service';
 import { CreateGuestDto } from './dto/create-guest.dto';
 import { UpdateGuestDto } from './dto/update-guest.dto';
 
@@ -8,10 +11,18 @@ export class GuestService {
   constructor(private readonly prisma: PrismaService) { }
 
   async create(createGuestDto: CreateGuestDto) {
-    const createdGuest = await this.prisma.guest.create({
-      data: { ...createGuestDto }
-    });
-    return createdGuest;
+    try {
+      const createdGuest = await this.prisma.guest.create({
+        data: { ...createGuestDto }
+      });
+      return createdGuest;
+    } catch (error) {
+      if (error instanceof PrismaClientKnownRequestError) {
+        if (error.code === 'P2005') {
+          throw new HttpException('invalid type in a field', HttpStatus.BAD_REQUEST);
+        }
+      }
+    }
   }
 
   async findAll() {
@@ -23,10 +34,15 @@ export class GuestService {
     const oneGuest = await this.prisma.guest.findUnique({
       where: { guest_email }
     });
+    if (!oneGuest) throw new HttpException('guest doesn\'t exist', HttpStatus.NOT_FOUND)
     return oneGuest;
   }
 
   async update(guest_email: string, updateGuestDto: UpdateGuestDto) {
+    const guestToBeUpdated = await this.prisma.guest.findUnique({
+      where: { guest_email }
+    });
+    if (!guestToBeUpdated) throw new HttpException('guest doesn\'t exist', HttpStatus.NOT_FOUND);
     const updatedUser = await this.prisma.guest.update({
       where: { guest_email },
       data: { ...updateGuestDto }
@@ -34,9 +50,13 @@ export class GuestService {
     return updatedUser;
   }
 
-  async remove(email: string) {
+  async remove(guest_email: string) {
+    const guestToBeDeleted = await this.prisma.guest.findUnique({
+      where: { guest_email }
+    });
+    if (!guestToBeDeleted) throw new HttpException('user doesn\'t exist', HttpStatus.NOT_FOUND)
     const deletedGuest = await this.prisma.guest.delete({
-      where: { guest_email: email }
+      where: { guest_email }
     });
     return deletedGuest;
   }
